Solved logout issues for db clean and credentials clear

+++ LocMessClient/app/src/main/java/pt/ulisboa/tecnico/cmu/tg14/locmessclient/Utils/FeedReaderDbHelper.java
@@ -134,8 +134,17 @@ public class FeedReaderDbHelper extends SQLiteOpenHelper {
 
     private static final String SQL_DELETE_ENTRIES = "DROP TABLE IF EXISTS ";
 
+    private static FeedReaderDbHelper mInstance = null;
+
     public FeedReaderDbHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
+    }
+
+    public static FeedReaderDbHelper getInstance(Context context){
+        if (mInstance == null) {
+            mInstance = new FeedReaderDbHelper(context.getApplicationContext());
+        }
+        return mInstance;
     }
 
     public void onCreate(SQLiteDatabase db) {
@@ -190,8 +199,8 @@ public class FeedReaderDbHelper extends SQLiteOpenHelper {
 
     public void dropLocation() {
         SQLiteDatabase db = this.getWritableDatabase();
-
-        db.execSQL(SQL_DELETE_ENTRIES + FeedEntry.LOCATION_TABLE_NAME);
+        if(db.isReadOnly())
+            db.execSQL(SQL_DELETE_ENTRIES + FeedEntry.LOCATION_TABLE_NAME);
     }
 
     public void deleteAllLocations() {
@@ -209,7 +218,8 @@ public class FeedReaderDbHelper extends SQLiteOpenHelper {
         contentValues.put(FeedEntry.LOCATION_COLUMN_LON, lon);
         contentValues.put(FeedEntry.LOCATION_COLUMN_RAD, radius);
         contentValues.put(FeedEntry.LOCATION_COLUMN_CENTRALIZED, centralized);
-        db.insert(FeedEntry.LOCATION_TABLE_NAME, null, contentValues);
+        if(!db.isReadOnly())
+            db.insert(FeedEntry.LOCATION_TABLE_NAME, null, contentValues);
         //Log.d("insertAllLocations: ","added to DB location " + name + ssid + ble + lat + lon + radius);
     }
 
@@ -760,9 +770,8 @@ public class FeedReaderDbHelper extends SQLiteOpenHelper {
         contentValues.put(FeedEntry.MESSAGE_COLUMN_ADDEDDECENTRALIZED, "false");
         contentValues.put(FeedEntry.MESSAGE_COLUMN_DELETEDDECENTRALIZED, "false");
         contentValues.put(FeedEntry.MESSAGE_COLUMN_NEARBY, "true");
-
-
-        db.insert(FeedEntry.MESSAGE_TABLE_NAME, null, contentValues);
+        if(!db.isReadOnly())
+            db.insert(FeedEntry.MESSAGE_TABLE_NAME, null, contentValues);
     }
 
     public void updateMessageInsertedToServer(String uuid) throws MultipleRowsAfectedException, MessageNotFoundException {
@@ -985,7 +994,9 @@ public class FeedReaderDbHelper extends SQLiteOpenHelper {
         String whereClause = FeedEntry.MESSAGE_COLUMN_NEARBY + " = ?";
         String[] whereArgs = new String[] { "true" };
 
-        int removed = db.delete(table, whereClause, whereArgs);
+        int removed = 0;
+        if(!db.isReadOnly())
+            removed = db.delete(table, whereClause, whereArgs);
 
         //Log.d("DBService", "deleteAllNearbyMessages -> " + removed);
 
@@ -1465,8 +1476,8 @@ public class FeedReaderDbHelper extends SQLiteOpenHelper {
 
     public void deleteAllServerProfiles() {
         SQLiteDatabase db = this.getWritableDatabase();
-
-        db.execSQL("delete from "+ FeedEntry.SERVER_PROFILES_TABLE_NAME);
+        if(!db.isReadOnly())
+            db.execSQL("delete from "+ FeedEntry.SERVER_PROFILES_TABLE_NAME);
     }
 
     private Profile associateServerProfiles(Cursor cursor) {
@@ -1596,4 +1607,13 @@ public class FeedReaderDbHelper extends SQLiteOpenHelper {
         db.execSQL("delete from "+ FeedEntry.MULE_PROFILE_TABLE_NAME);
     }
 
+    public void deleteAll() {
+        this.deleteAllMessages();
+        this.deleteAllLocations();
+        this.deleteAllMessageMules();
+        this.deleteAllProfiles();
+        this.deleteAllMuleProfiles();
+        this.deleteAllServerProfiles();
+        this.deleteAllNearbyMessages();
+    }
 }
